Update README.md
The C++ program, CGPA Calculator, is designed to determine a student’s cumulative grade point average (CGPA). By using the user’s input, which includes details such as the number of courses taken and the grades earned in each, the program computes the CGPA. Additionally, it displays each student’s overall course grade. The CGPA Calculator processes the given exam results to calculate the CGPA, showing individual course grades, calculating total credits and grade points achieved, and determining the GPA for the semester. Based on all this data, it generates and presents the student’s CGPA.